+++ client/src/components/CommunityInviteBox/index.tsx
@@ -4,7 +4,7 @@ import Button from '@components/Button';
 import CommunityInviteUserSearchResult from '@components/CommunityInviteUserSearchResult';
 import ErrorMessage from '@components/ErrorMessage';
 import SearchInput from '@components/SearchInput';
-import useUsersQuery from '@hooks/useUsersQuery';
+import { useUsersQuery } from '@hooks/user';
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
